requirements.txt
astroid==1.4.9
autoflake==0.6.6
autopep8==1.2.4
backports.functools-lru-cache==1.3
cloudinary==1.6.0
configparser==3.5.0
coverage==4.2
dill==0.2.6
dj-places==1.1.1
Django==1.9.2
django-allauth==0.29.0
django-extensions==1.7.6
django-filter==1.0.1
django-jenkins==0.110.0
django-rest-auth==0.9.0
djangorestframework==3.5.0
dropbox==7.1.1
ecdsa==0.13
enum34==1.1.6
extras==1.0.0
fixtures==3.0.0
flake8==3.2.1
funcsigs==1.0.2
future==0.16.0
futures==3.0.5
fuzzywuzzy==0.14.0
gapic-google-cloud-datastore-v1==0.14.1
gapic-google-cloud-logging-v2==0.90.1
gapic-google-cloud-pubsub-v1==0.14.1
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
google-api-python-client==1.6.2
google-auth==0.6.0
google-auth-httplib2==0.0.2
google-cloud==0.22.0
google-cloud-bigquery==0.22.1
google-cloud-bigtable==0.22.0
google-cloud-core==0.22.1
google-cloud-datastore==0.22.1
google-cloud-dns==0.22.0
google-cloud-error-reporting==0.22.0
google-cloud-happybase==0.22.0
google-cloud-language==0.22.2
google-cloud-logging==0.22.0
google-cloud-monitoring==0.22.0
google-cloud-pubsub==0.22.0
google-cloud-resource-manager==0.22.0
google-cloud-runtimeconfig==0.22.0
google-cloud-storage==0.22.0
google-cloud-translate==0.22.0
google-cloud-vision==0.22.0
google-gax==0.15.4
googleapis-common-protos==1.5.2
GoogleAppEngineCloudStorageClient==1.9.22.1
grpc-google-cloud-datastore-v1==0.14.0
grpc-google-cloud-logging-v2==0.90.0
grpc-google-cloud-pubsub-v1==0.14.0
grpc-google-iam-v1==0.11.1
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.1.0
gunicorn==19.6.0
httplib2==0.10.3
isort==4.2.5
jws==0.1.3
lazy-object-proxy==1.2.2
lettuce==0.2.23
linecache2==1.0.0
Markdown==2.6.7
mccabe==0.5.3
mechanize==0.2.5
mock==2.0.0
nose==1.3.7
oauth2client==3.0.0
oauthlib==2.0.1
pbr==1.10.0
pep8==1.7.0
ply==3.8
protobuf==3.2.0
psycopg2==2.6.2
pyasn1==0.2.2
pyasn1-modules==0.0.8
pycodestyle==2.2.0
pycrypto==2.6.1
pycryptodome==3.4.3
PyDrive==1.3.1
pyflakes==1.3.0
pylint==1.6.4
pylint-django==0.7.2
pylint-plugin-utils==0.2.4
Pyrebase==3.0.27
python-jose==1.3.2
python-jwt==2.0.1
python-Levenshtein==0.12.0
python-mimeparse==1.6.0
python-openid==2.2.5
python-subunit==1.2.0
PyYAML==3.12
requests==2.12.4
requests-oauthlib==0.7.0
requests-toolbelt==0.7.0
rsa==3.4.2
setuptools-sloccount==0.1
six==1.10.0
sure==1.4.0
testtools==2.2.0
traceback2==1.4.0
typing==3.5.3.0
unittest2==1.1.0
uritemplate==3.0.0
urllib3==1.20
virtualenv==15.1.0
wheel==0.24.0
whitenoise==3.2.2
wrapt==1.10.8
